Refactor benchmark framework and add CDC E2E test
- Move benchmarks from tests/ to bench/ with auto-discovery orchestrator
- Simplify justfile from 6 bench targets to 2: `just bench` and `just bench-compare`
- Add per-connector config/setup/run/teardown pattern (bench/connectors/postgres/)
- Add criterion-style statistical report (confidence intervals, throughput)
- Add Criterion 0.5 micro-benchmarks for arrow_codec and state_backend
- Update docs/BENCHMARKING.md with new structure and commands
- Fix CDC bug: CDC_MAX_CHANGES was i64 but pg_logical_slot_get_changes expects int4
- Add CDC E2E test (15_cdc.sh) exercising INSERT/UPDATE/DELETE capture
- Enable wal_level=logical in docker-compose for CDC support

+++ connectors/source-postgres/src/cdc.rs
@@ -27,7 +27,8 @@ use crate::schema::pg_type_to_arrow;
 const BATCH_SIZE: usize = 10_000;
 
 /// Maximum WAL changes consumed per CDC invocation to avoid unbounded memory use.
-const CDC_MAX_CHANGES: i64 = 100_000;
+/// Type is i32 because pg_logical_slot_get_changes() expects int4.
+const CDC_MAX_CHANGES: i32 = 100_000;
 
 /// Default replication slot prefix. Full slot names are `rapidbyte_{stream_name}`.
 const SLOT_PREFIX: &str = "rapidbyte_";
